+++ Projects/STM32F429I-Discovery/Examples/MY_LIB/gpio_c1.c
@@ -81,7 +81,7 @@ static int gpio_c1_read(void)
 	return HAL_GPIO_ReadPin(PC1_PORT, PC1_PIN);
 }
 
-static void gpio_c1_write(uint16_t state)
+static void gpio_c1_write(uint8_t state)
 {
 	if(state == 0) {
 		HAL_GPIO_WritePin(PC1_PORT, PC1_PIN, GPIO_PIN_RESET);
